app/requirements.txt
streamlit==1.24.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-api-python-client==2.88.0
pandas==2.0.2
python-dotenv==1.0.0 